[WIP] After move from react-widget, this routes util isn't working so well
On init the URLs get interpolated (but with missing data), then we try to
interpolate them again with new data... but there are no more `{}` left.
So we get `undefined` all over the shop.

You know what, it's because `format` is implemented differently in HUI
than it is in react widgets. Ok, I can sort that out.