Ensures the call to setState made in addressAPI callback is made second
When setState is called in this method and addressAPI.search() is async,
everything is a-ok. If the call is synchronous then the callback passed
to it is executed before the setState it's inside. So we'd set the
addressList in the callback, and then the setState wrapping around would
set it back to `null`... good times.