Accounts for different browser behaviour for FilterSelect
iOS and OSX require use of a proper form element for capturing focus.
IE: onmousedown opens the select dropdown, need to act on that event
rather than click.
iOS: ignores preventDefault on focus of selects, so the display select
needs to "work" when the input has been tabbed to. This is _very_
annoying and less than ideal.