Adds test for requireValue call on blur
* Need to figure out a way to test the debouncing...

The debounce is there because we want to know if no input inside the
component has focus... when one input is blurring and another receiving
focus the progression will be BLUR > FOCUS so we need blur events with
no following focus to know that the entire component is blurred.

TODO: Good candidate for a reactive stream!