Adds aggregateBlurrable mixin
When a component has many elements within it which can recieve focus, we
need a way of knowing if the "whole" component is focused or blurring.

The answer to question "is it focused?" is easier than "is it blurred",
the reason is that there is a time gap between the blurring of one child
element and focussing of another. An aggregate blur is therefore when
one child blurs and no other gets focus within an acceptable time frame.

It ain't perfect but with this mixin we can pop "handleBlur" and
"handleFocus" on any child with a component and be able to know whether
the whole component has blurred with the "handleAggregateBlur" method.