Allow touch scrolling the option list
We can't use click, as it fires after blur and we use blur to close the
options list in the filter select / country select / url search select

So here we use mousedown / touch start, touchstart fires when the user
tries to scroll. To get around that triggering a selection, we keep a
var which tells us whether or not the touch has moved, if it's moved
selection wasn't intended so don't select.