Only pluck valid and messages properties if onError is defined
This still isn't the greatest thing in the world but I got bitten by the destructuring here. The built in assumption that getValidator would return an object with 'valid' and 'messages' properties is probably fair if you are using the onError prop, but otherwise it caused some headaches. Hopefully this is a reasonable compromise.