Scope blurring timer in FocusAggregate per instance
I'm not able to recreate them, but we have a few errors in sentry that
_might_ be related to methods being called on objects which have been
"cleaned up" ready for reuse. I don't know if this is something react
does. The idea is to try and make this a little more reasonable (it's
not totally reasonable because the job it's doing is pretty weird). I
doubt this will have any effect of on those errors ... but if it does:
super.